requirements.txt
streamlit==1.37.1
tensorflow==2.16.1
Pillow==10.4.0
scikit-learn==1.5.1
numpy==1.26.4
protobuf==3.20.*
typing-extensions>=4.5.0


